👌 IMPROVE: Improve navbar dir structure
