+++ src/components/sidebar/SidebarHorizontal.jsx
@@ -66,7 +66,6 @@ const Topbar = () => {
         icon={<ContactSupport size="22" />}
         to="/dashboard/contact"
       /> */}
-
     </Container>
   )
 }
